Remove main.py
